DTA <--> DTB conversion
Added an option to ark unpack to convert dtb files to dta

During pack, any dta files of the following extensions will be converted to dtb if there isn't already a dtb version available.

- dta
- fusion
- moggsong
- script

It should be noted that the underlying DtxCS library used for this conversion cannot successfully convert all of the dta files back to dtb files, it is recommended that the option keep the dtb files is chosen.  This allows for selectively deleting the DTB to trigger a build of the modified DTA during pack.