Parse doubles with InvariantCulture
By default, `double.Parse()` handles decimal points based on the user's system preference for region in Windows.

When the user's region is set to something that prefers decimal points other than `.`, a `FormatException` is thrown by `double.Parse()`, which we silently ignore and default values to 0 instead, as per the parsing code in `MoggSong`.

This can be mitigated by using `CultureInfo.InvariantCulture` as the format provider for `double.Parse()`